+++ _metadata.py
@@ -1,2 +1,2 @@
-__extension_version__ = "0.50.0"
+__extension_version__ = "0.51.0rc0"
 __extension_name__ = "pytket-qiskit"