+++ docs/changelog.rst
@@ -5,6 +5,8 @@ Unreleased
 ----------
 
 * Update qiskit-ibm-runtime version requirement to 0.22.
+* remove all remote simulators
+* rename ``IBMQLocalEmulatorBackend`` to ``IBMQEmulatorBackend``
 
 0.50.0 (March 2024)
 -------------------